requirements.txt
ddddocr~=1.5.6
pillow~=11.1.0